separate femlearn into separate modules
