+++ blueprints/utils/femlearn/loads.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from blueprints.utils.femlearn import _LoadsBaseClass
+from blueprints.utils.femlearn._base_classes import _LoadsBaseClass
 
 
 class LoadsOnNodes(_LoadsBaseClass):
